requirements.txt
# requirements.txt
# Core Framework dan Server
fastapi
uvicorn

# Database dan ORM
sqlalchemy
pydantic

# Keamanan (JWT dan Hashing Password)
python-jose[cryptography] 
passlib[argon2]          

# Utilitas Tambahan
python-multipart

uvicorn
gunicorn